Fix placeholder lifetime bug in memory planning (#6971)
Summary:

If we have a node that is inserted between placeholders via a pass after to_edge then there is a bug in memory planning lifetime calculations which results in two placeholders being allocated the same memory segment.

Two placeholders should never be using the same memory segment and to prevent this we set the beginning lifetime of the placeholder to be always 0 in the memory planning pass.

In the test case added this is the graph:
```
graph():
    %a : [num_users=1] = placeholder[target=a]
    %aten_permute_copy_default : [num_users=1] = call_function[target=executorch.exir.dialects.edge._ops.aten.permute_copy.default](args = (%a, [1, 0, 2]), kwargs = {})
    %b : [num_users=2] = placeholder[target=b]
    %aten_add_tensor : [num_users=2] = call_function[target=executorch.exir.dialects.edge._ops.aten.add.Tensor](args = (%aten_permute_copy_default, %b), kwargs = {})
    %aten_add_tensor_1 : [num_users=1] = call_function[target=executorch.exir.dialects.edge._ops.aten.add.Tensor](args = (%aten_add_tensor, %b), kwargs = {})
    return (aten_add_tensor, aten_add_tensor_1)
```

Without this fix the lifetimes of placeholders a and b are:
`a => [0, 2]`
`b => [3, 7]`
Thus the same memory segment is allocated for both of them.

After this fix the lifetimes of the placeholders a and b are:
`a => [0, 2]`
`b => [0, 7]`

Differential Revision: D66184849